pyproject.toml
[tool.poetry]
name = "baseplate"
version = "2.7.1b4"
description = "reddit's python service framework"
authors = ["reddit"]
license = "BSD"
readme = "README.md"
homepage = "https://github.com/reddit/baseplate.py"
repository = "https://github.com/reddit/baseplate.py"
documentation = "https://baseplate.readthedocs.io/en/stable/"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: BSD License",
    "Operating System :: POSIX :: Linux",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
advocate = { version = ">=1.0.0,<2.0", optional = true }
boto3 = ">=1.28.27"
cassandra-driver = { version = ">=3.29.0,<4.0", optional = true }
confluent-kafka = { version = ">=2.3.0", optional = true }
gevent = ">=23.9.1"
kazoo = { version = ">=2.5.0,<3.0", optional = true }
kombu = { version = ">=5.3.3", optional = true }
objgraph = { version = ">=3.6.0", optional = true }
posix-ipc = ">=1.0.0,<2.0"
prometheus-client = ">=0.12.0"
pymemcache = { version = ">=1.3.0,<1.4.4", optional = true }
pyramid = { version = ">=1.10.8,<2.0", optional = true }
python-json-logger = ">=2.0.5,<3.0"
reddit-cqlmapper = { version = ">=0.3.0,<1.0", optional = true }
redis = { version = ">=2.10.0,<4.0.0", optional = true }
redis-py-cluster = { version = ">=2.1.2,<3.0.0", optional = true }
requests = ">=2.21.0,<3.0"
sentry-sdk = { version = ">=1.35.0,<2.0", optional = true }
sqlalchemy = { version = ">=1.4.49,<2", optional = true }
thrift-unofficial = ">=0.19.0,<1.0"
typing-extensions = "^4.11.0"
opentelemetry-sdk = "^1.26.0"
opentelemetry-api = "^1.26.0"
opentelemetry-instrumentation-pyramid = "^0.47b0"
opentelemetry-instrumentation-requests = "^0.47b0"
opentelemetry-instrumentation-threading = "^0.47b0"
opentelemetry-instrumentation-logging = "^0.47b0"
opentelemetry-exporter-otlp = "^1.26.0"
formenergy-observability = "^0.3.2"
pyrate-limiter = "^3.6.1"
psycopg2 = "^2.0.0"
psycogreen = "^1.0.0"

[tool.poetry.extras]
amqp = ["kombu"]
cassandra = ["cassandra-driver"]
cqlmapper = ["reddit-cqlmapper"]
kafka = ["confluent-kafka"]
memcache = ["pymemcache"]
prometheus = []  # Kept for backwards compatibility, these are now main requirements
pyramid = ["pyramid"]
redis = ["redis"]
redis-py-cluster = ["redis-py-cluster"]
refcycle = ["objgraph"]
requests = ["advocate"]
s3fetcher = []  # Kept for backwards compatibility, these are now main requirements
sentry = ["sentry-sdk"]
sql = ["sqlalchemy"]
psycopg2 = ["psycopg2", "psycogreen"]
zookeeper = ["kazoo"]

[tool.poetry.group.dev.dependencies]
fakeredis = "*"
lxml = "*"
moto = "*"
mypy = "*"
pre-commit = "*"
pydocstyle = "*"
pylint = "*"
pytest = "7.4.4"
pytest-cov = "*"
pytz = "*"
sphinx = "*"
sphinx-autodoc-typehints = "*"
types-redis = "*"
types-requests = "*"
types-setuptools = "*"
webtest = "*"
parameterized = "^0.9.0"
opentelemetry-test-utils = "^0.47b0"
ruff = "*"


[tool.poetry.scripts]
baseplate-healthcheck = { reference = "bin/baseplate-healthcheck", type = "file" }
baseplate-script = { reference = "bin/baseplate-script", type = "file" }
baseplate-serve = { reference = "bin/baseplate-serve", type = "file" }
baseplate-shell = { reference = "bin/baseplate-shell", type = "file" }
baseplate-tshell = { reference = "bin/baseplate-tshell", type = "file" }

[tool.poetry.plugins."distutils.commands"]
build_thrift = "baseplate.frameworks.thrift.command:BuildThriftCommand"

[tool.ruff]
target-version = "py39"
line-length = 100
extend-exclude = ["baseplate/thrift", "tests/integration/test_thrift"]

[tool.ruff.lint]
extend-select = [
  "I",  # isort
  "UP",  # pyupgrade
  "E501", # line length
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
